refactor: make modules for retriable_*_clients
